Declaring commit bankruptcy and lumping changes
* Add support for reading/writing float or double matrices, not just
  uint32_t, both in R and C++
* Add initial functionality to allow transposing matrices with
  an external merge sort for memory scalability
* Add fragment rename cells option, which sadly required changing the
  `*Name` options to no longer be const
* Create initial functionality preparing for ATAC geneScore calculations
* Create draft functionality for reading BP128 arrays in R
* Bugfix in seeking BP128 arrays

+++ src/matrixIterators/ImportMatrixHDF5.h
@@ -31,7 +31,7 @@ StoredMatrix<uint32_t> open10xFeatureMatrix(std::string file, uint32_t buffer_si
 
 // Read AnnData sparse matrix, with an implicit transpose to CSC format for
 // any data stored in CSR format
-StoredMatrix<double> openAnnDataMatrix(std::string file, std::string group, uint32_t buffer_size, uint32_t read_size=1024);
+StoredMatrix<float> openAnnDataMatrix(std::string file, std::string group, uint32_t buffer_size, uint32_t read_size=1024);
 
 bool isRowOrientedAnnDataMatrix(std::string file, std::string group);
